jobs/embed_call/requirements.txt
flask
gunicorn
google-cloud-bigquery
google-cloud-pubsub
google-cloud-storage
google-cloud-aiplatform 
vertexai>=1.44.0
pandas